# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Text-to-SQL & Document Chat Platform",
    "description": "Multi-tenant platform for natural language database queries and document chat",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/database-connections": {
      "post": {
        "tags": [
          "connections"
        ],
        "summary": "Create Connection",
        "description": "Create a new database connection. Password is encrypted at rest.",
        "operationId": "create_connection_api_database_connections_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "connections"
        ],
        "summary": "List Connections",
        "description": "List all database connections for the current tenant.",
        "operationId": "list_connections_api_database_connections_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionOut"
                  },
                  "title": "Response List Connections Api Database Connections Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/database-connections/{connection_id}": {
      "get": {
        "tags": [
          "connections"
        ],
        "summary": "Get Connection",
        "description": "Get a specific database connection by ID.",
        "operationId": "get_connection_api_database_connections__connection_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "connections"
        ],
        "summary": "Update Connection",
        "description": "Update a database connection. Re-encrypts password if changed.",
        "operationId": "update_connection_api_database_connections__connection_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Connection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "connections"
        ],
        "summary": "Delete Connection",
        "description": "Delete a database connection.",
        "operationId": "delete_connection_api_database_connections__connection_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/database-connections/{connection_id}/test": {
      "post": {
        "tags": [
          "connections"
        ],
        "summary": "Test Connection Endpoint",
        "description": "Test a database connection. Decrypts credentials transiently in memory.",
        "operationId": "test_connection_endpoint_api_database_connections__connection_id__test_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/permissions/connections/{connection_id}": {
      "get": {
        "tags": [
          "permissions"
        ],
        "summary": "List Connection Permissions",
        "description": "List configured table permissions for a connection.",
        "operationId": "list_connection_permissions_api_permissions_connections__connection_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TablePermissionOut"
                  },
                  "title": "Response List Connection Permissions Api Permissions Connections  Connection Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ColumnPermissionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "table_permission_id": {
            "type": "string",
            "format": "uuid",
            "title": "Table Permission Id"
          },
          "role_id": {
            "type": "string",
            "format": "uuid",
            "title": "Role Id"
          },
          "column_name": {
            "type": "string",
            "title": "Column Name"
          },
          "is_allowed": {
            "type": "boolean",
            "title": "Is Allowed"
          },
          "is_masked": {
            "type": "boolean",
            "title": "Is Masked"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "table_permission_id",
          "role_id",
          "column_name",
          "is_allowed",
          "is_masked"
        ],
        "title": "ColumnPermissionOut",
        "description": "Output model for column permission."
      },
      "ConnectionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "database_type": {
            "type": "string",
            "pattern": "^(postgresql|mysql)$",
            "title": "Database Type"
          },
          "host": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port"
          },
          "database_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Database Name"
          },
          "username": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password"
          },
          "ssl_enabled": {
            "type": "boolean",
            "title": "Ssl Enabled",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "database_type",
          "host",
          "port",
          "database_name",
          "username",
          "password"
        ],
        "title": "ConnectionCreate",
        "description": "Create a new database connection."
      },
      "ConnectionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "database_type": {
            "type": "string",
            "title": "Database Type"
          },
          "host": {
            "type": "string",
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "title": "Port"
          },
          "database_name": {
            "type": "string",
            "title": "Database Name"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "ssl_enabled": {
            "type": "boolean",
            "title": "Ssl Enabled"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "name",
          "database_type",
          "host",
          "port",
          "database_name",
          "username",
          "ssl_enabled",
          "is_active"
        ],
        "title": "ConnectionOut",
        "description": "Connection data returned to the client.\n\nNote: password is NEVER returned \u2014 only the encrypted_password column\nexists in the DB, and we never expose that either."
      },
      "ConnectionUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "host": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 65535.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "database_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "ssl_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Enabled"
          }
        },
        "type": "object",
        "title": "ConnectionUpdate",
        "description": "Update an existing database connection."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TablePermissionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "role_id": {
            "type": "string",
            "format": "uuid",
            "title": "Role Id"
          },
          "connection_id": {
            "type": "string",
            "format": "uuid",
            "title": "Connection Id"
          },
          "schema_name": {
            "type": "string",
            "title": "Schema Name"
          },
          "table_name": {
            "type": "string",
            "title": "Table Name"
          },
          "access_type": {
            "type": "string",
            "title": "Access Type"
          },
          "row_filter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Filter"
          },
          "column_permissions": {
            "items": {
              "$ref": "#/components/schemas/ColumnPermissionOut"
            },
            "type": "array",
            "title": "Column Permissions",
            "default": []
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role_id",
          "connection_id",
          "schema_name",
          "table_name",
          "access_type"
        ],
        "title": "TablePermissionOut",
        "description": "Output model for table permission."
      },
      "TestResult": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "TestResult",
        "description": "Result of a connection test."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}